incorporate req'd Kotlin files
because Android Studio decided not to, despite me telling the program to incorporate a kotlin file. What did it think I wanted? Just create it and not be able to read it? :facepalm: so I created a new kotlin project, that *did* have the necessary gradle build files for me to reference.